jira: MMBOB-37 when a user logs in, the 'seen' flag is put to 'false'. so the first time 'signalSeen()' is called, the user is put on the sync que. So now the user profile is actually updated to the remote profile.
When the forum starts up now it is made certain that there is a profile entry for each profile entry def in the configuration. if it is not there, and the profileinfo is synced, only the profile entries that are in the mmbob profile are synced and that is wrong. all fields should be synced.